added preventing manipulation of the checkins that are 7 days old

+++ routes/checkins.js
@@ -1,6 +1,8 @@
 'use strict';
 
 var uuid = require('uuid');
+var moment = require('moment');
+
 var Utils      = require('../utils/utils');
 
 var Checkin    = require('../models/checkin');
@@ -49,24 +51,32 @@ module.exports = require('koa-router')()
     this.response.status = 400;
     this.body = { "error" : 'parameters missing'};
   } else {
-    var employee =
-      new Employee({ activation_code: this.params.activation_code});
-    employee.loadByActivationCode();
+    let momentNow = moment(new Date());
+    let momentCheckedInAt = moment(checked_in_at);
+    var duration = moment.duration(momentNow.diff(momentCheckedInAt));
+    var days = duration.asDays();
 
-    var checkin = new Checkin(
-      {
-        email: employee.email,
-        user_id: employee.user_id,
-        checked_in_at: checked_in_at,
-        duration: 0,//always create with duration 0, can modify later by updating checkin
-        action_code_id: action_code_id
-      }
-    );
+    if(days < 0 || days > 7) {
+      this.response.status = 403;
+      this.body = { "error" : 'unable to create more then 7 days old checkin'};
+    } else {
+      var employee =
+        new Employee({ activation_code: this.params.activation_code});
+      employee.loadByActivationCode();
 
-    // console.log(checkin);
-    checkin.save();
-    this.response.status = 200;
-    this.body = { "checkin" : checkin };
+      var checkin = new Checkin(
+        {
+          email: employee.email,
+          user_id: employee.user_id,
+          checked_in_at: checked_in_at,
+          duration: 0,//always create with duration 0, can modify later by updating checkin
+          action_code_id: action_code_id
+        }
+      );
+      checkin.save();
+      this.response.status = 200;
+      this.body = { "checkin" : checkin };
+    }
   }
 })
 
@@ -90,7 +100,6 @@ module.exports = require('koa-router')()
       this.response.status = 200;
       this.body = { "checkin" : checkin };
     }
-
 })
 
 //update checkin which includes checkout, duration must be passed as a parameter
@@ -110,20 +119,40 @@ module.exports = require('koa-router')()
     this.response.status = 404;
     this.body = { "error" : 'checkin not found' };
   } else {
-    let data = this.request.body;
+    let momentNow = moment(new Date());
+    let momentCheckedInAt = moment(checkin.checked_in_at);
+    var duration = moment.duration(momentNow.diff(momentCheckedInAt));
+    var days = duration.asDays();
 
-    let checked_in_at = data.checked_in_at;
-    let duration = data.duration;
-    let action_code_id = data.action_code_id;
+    if(days > 7) {
+      this.response.status = 403;
+      this.body = { "error" : 'unable to update more then 7 days old checkin'};
+    } else {
 
-    checkin.checked_in_at = checked_in_at;
-    checkin.duration = duration;
-    checkin.action_code_id = action_code_id;
+      let data = this.request.body;
 
-    checkin.save();
+      let checked_in_at = data.checked_in_at;
+      let duration = data.duration;
+      let action_code_id = data.action_code_id;
 
-    this.response.status = 200;
-    this.body = { "checkin" : checkin };
+      checkin.checked_in_at = checked_in_at;
+      checkin.duration = duration;
+      checkin.action_code_id = action_code_id;
+
+      momentCheckedInAt = moment(checkin.checked_in_at);
+      duration = moment.duration(momentNow.diff(momentCheckedInAt));
+      days = duration.asDays();
+
+      if(days < 0 || days > 7) {
+        this.response.status = 403;
+        this.body = { "error" : 'unable to update to more then 7 days old checkin'};
+      } else {
+        checkin.save();
+
+        this.response.status = 200;
+        this.body = { "checkin" : checkin };
+      }
+    }
   }
 })
 
@@ -144,9 +173,20 @@ module.exports = require('koa-router')()
     this.response.status = 404;
     this.body = { "error" : 'checkin not found' };
   } else {
-    checkin.delete();
-    this.response.status = 200;
-    this.body = { "result" : 'checkin deleted' };
+
+    let momentNow = moment(new Date());
+    let momentCheckedInAt = moment(checkin.checked_in_at);
+    var duration = moment.duration(momentNow.diff(momentCheckedInAt));
+    var days = duration.asDays();
+
+    if(days > 7) {
+      this.response.status = 403;
+      this.body = { "error" : 'unable to delete more then 7 days old checkin'};
+    } else {
+      checkin.delete();
+      this.response.status = 200;
+      this.body = { "result" : 'checkin deleted' };
+    }
   }
 })
